Extract methods to AggregationMixin::Methods
